feat: add new dynamic on click yes risc value
