+++ src/App.js
@@ -5,7 +5,6 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "bootstrap/dist/css/bootstrap.min.css";
 import Patients from './components/Patients';
 import Details from './components/Details';
-import DetailsV2 from './components/DetailsV2';
 
 function App() {
   return (
@@ -14,7 +13,7 @@ function App() {
         <Routes>
           <Route exact path={"/"} element={<Home />} />
           <Route exact path={"/patients/:crm"} element={<Patients />} />
-          <Route exact path={"/details/v2/:crm"} element={<DetailsV2 />} />
+          <Route exact path={"/details/v2/:crm"} element={<Details />} />
         </Routes>
       </Router>
     </div>
