+++ src/constants/defaultObjects.js
@@ -1,0 +1,124 @@
+export const defaultObject = {
+    id: "",
+    hepatiteB: false,
+    crm: "",
+    hepatiteB_obs: "",
+    relevant_hepatiteB: false,
+    risco: 0,
+    risco_sim: null,
+    risco_nao: null,
+    obs_geral: "",
+    pessoa: "",
+    pessoa_obs: "",
+    pessoa_eh_relevante: false,
+    idade: "",
+    idade_obs: "",
+    idade_eh_relevante: "",
+    sexo: "",
+    sexo_obs: "",
+    sexo_eh_relevante: false,
+    data: "",
+    data_obs: "",
+    data_eh_relevante: false,
+    exame: "",
+    exame_obs: "",
+    exame_eh_relevante: false,
+    valor_obtido_em_exame: "",
+    valor_obtido_em_exame_obs: "",
+    valor_obtido_em_exame_eh_relevante: false,
+    condicao_medica: "",
+    condicao_medica_obs: "",
+    condicao_medica_eh_relevante: "",
+    tuberculose: false,
+    tuberculose_obs: "",
+    tuberculose_eh_relevante: false,
+    caxumba: false,
+    caxumba_obs: "",
+    caxumba_eh_relevante: false,
+    sarampo: false,
+    sarampo_obs: "",
+    sarampo_eh_relevante: false,
+    hepatite_A: false,
+    hepatite_A_obs: "",
+    hepatite_A_eh_relevante: false,
+    rubeola: false,
+    rubeola_obs: "",
+    rubéola_eh_relevante: false,
+    herpes_zoster: false,
+    herpes_zoster_obs: "",
+    herpes_zoster_eh_relevante: false,
+    varicela: false,
+    varicela_obs: "",
+    varicela_eh_relevante: false,
+    covid_19: false,
+    covid_19_obs: "",
+    covid_19_eh_relevante: false,
+    historico_de_cirurgias: false,
+    historico_de_cirurgias_obs: "",
+    historico_de_cirurgias_eh_relevante: false,
+    numero_de_gravidezes: "",
+    numero_de_gravidezes_obs: "",
+    numero_de_gravidezes_eh_relevante: false,
+    numero_de_partos: "",
+    numero_de_partos_obs: "",
+    numero_de_partos_eh_relevante: "",
+    numero_de_abortos: "",
+    numero_de_abortos_obs: "",
+    numero_de_abortos_eh_relevante: "",
+    historico_de_alergias: false,
+    historico_de_alergias_obs: "",
+    historico_de_alergias_eh_relevante: false,
+    problema_de_audicao: false,
+    problema_de_audicao_obs: "",
+    problema_de_audicao_eh_relevante: false,
+    utiliza_medicacao_com_frequencia: false,
+    utiliza_medicacao_com_frequencia_obs: "",
+    utiliza_medicacao_com_frequencia_eh_relevante: false,
+    historico_de_gravidez: false,
+    historico_de_gravidez_obs: "",
+    historico_de_gravidez_eh_relevante: false,
+    pratica_exercicios_fisicos: false,
+    pratica_exercicios_fisicos_obs: "",
+    pratica_exercicios_fisicos_eh_relevante: false,
+    consumo_de_alcool: false,
+    consumo_de_alcool_obs: "",
+    consumo_de_alcool_eh_relevante: false,
+    fumante: "",
+    fumante_obs: "",
+    fumante_eh_relevante: "",
+    cigarros_por_dia: "",
+    cigarros_por_dia_obs: "",
+    cigarros_por_dia_eh_relevante: "",
+    anos_de_fumante: "",
+    anos_de_fumante_obs: "",
+    anos_de_fumante_eh_relevante: "",
+    historico_familiar_de_hipertensao: false,
+    historico_familiar_de_hipertensao_obs: "",
+    historico_familiar_de_hipertensao_eh_relevante: false,
+    historico_familiar_de_diabetes: false,
+    historico_familiar_de_diabetes_obs: "",
+    historico_familiar_de_diabetes_eh_relevante: false,
+    historico_familiar_de_cancer: false,
+    historico_familiar_de_cancer_obs: "",
+    historico_familiar_de_cancer_eh_relevante: false,
+    historico_familiar_de_doencas_neurologicas: false,
+    historico_familiar_de_doencas_neurologicas_obs: "",
+    historico_familiar_de_doencas_neurologicas_eh_relevante: false,
+    historico_familiar_de_infarto: "",
+    historico_familiar_de_infarto_obs: "",
+    historico_familiar_de_infarto_eh_relevante: "",
+    historico_familiar_de_infarto_agudo_do_miocardio: false,
+    historico_familiar_de_infarto_agudo_do_miocardio_obs: "",
+    historico_familiar_de_infarto_agudo_do_miocardio_eh_relevante: false,
+    historico_familiar_de_outras_doencas: "",
+    historico_familiar_de_outras_doencas_obs: "",
+    historico_familiar_de_outras_doencas_eh_relevante: "",
+    quais_valor_obtido_em_exame_sao_relevantes: "",
+    grau_certeza: null,
+    IMC_eh_relevante: false,
+    PRESSAO_ARTERIAL_eh_relevante: false,
+    PRESSAO_ARTERIAL_PAD_eh_relevante: false,
+    PRESSAO_ARTERIAL_PAD_eh_relevante: false,
+    utiliza_medicacao_com_frequencia_eh_relevante: false,
+    PRESSAO_ARTERIAL_PAS_eh_relevante: false,
+  };
